Fehler bei Datenrange von Datenbank Einträgen geändert

+++ accounting-service/src/main/resources/db/migration/V01.0001__init.sql
@@ -3,7 +3,7 @@ CREATE SCHEMA IF NOT EXISTS accountingservice;
 CREATE TABLE IF NOT EXISTS accountingservice.accounting
 (
     ID               SERIAL PRIMARY KEY,
-    PASSENGER        INTEGER,
+    PASSENGER        BIGINT,
     SEAT             VARCHAR(50),
     ACCOUNTINGSTATUS VARCHAR(25),
     UNIQUE (SEAT)
